requirements.txt
paho-mqtt
pymongo
scikit-learn
joblib
flask
pandas
config